_:construction: Refactor project model
